requirements.txt
pandas
pandas-ta
requests
numpy
python-binance
pybit